Split waveform audio file helpers
